fix bugs and apply yapf format

+++ models/pre_act_resnet.py
@@ -5,20 +5,28 @@ from torch.autograd import Variable
 import math
 from functools import partial
 
-__all__ = ['PreActivationResNet', 'resnet18', 'resnet34', 'resnet50', 'resnet101', 'resnet152', 'resnet200']
+__all__ = [
+    'PreActivationResNet', 'resnet18', 'resnet34', 'resnet50', 'resnet101',
+    'resnet152', 'resnet200'
+]
 
 
 def conv3x3x3(in_planes, out_planes, stride=1):
     # 3x3x3 convolution with padding
-    return nn.Conv3d(in_planes, out_planes, kernel_size=3,
-                     stride=stride, padding=1, bias=False)
+    return nn.Conv3d(
+        in_planes,
+        out_planes,
+        kernel_size=3,
+        stride=stride,
+        padding=1,
+        bias=False)
 
 
 def downsample_basic_block(x, planes, stride):
     out = F.avg_pool3d(x, kernel_size=1, stride=stride)
-    zero_pads = torch.Tensor(out.size(0), planes - out.size(1),
-                             out.size(2), out.size(3),
-                             out.size(4)).zero_()
+    zero_pads = torch.Tensor(
+        out.size(0), planes - out.size(1), out.size(2), out.size(3),
+        out.size(4)).zero_()
     if isinstance(out.data, torch.cuda.FloatTensor):
         zero_pads = zero_pads.cuda()
 
@@ -67,8 +75,8 @@ class PreActivationBottleneck(nn.Module):
         self.bn1 = nn.BatchNorm3d(inplanes)
         self.conv1 = nn.Conv3d(inplanes, planes, kernel_size=1, bias=False)
         self.bn2 = nn.BatchNorm3d(planes)
-        self.conv2 = nn.Conv3d(planes, planes, kernel_size=3, stride=stride,
-                               padding=1, bias=False)
+        self.conv2 = nn.Conv3d(
+            planes, planes, kernel_size=3, stride=stride, padding=1, bias=False)
         self.bn3 = nn.BatchNorm3d(planes)
         self.conv3 = nn.Conv3d(planes, planes * 4, kernel_size=1, bias=False)
         self.relu = nn.ReLU(inplace=True)
@@ -100,21 +108,36 @@ class PreActivationBottleneck(nn.Module):
 
 class PreActivationResNet(nn.Module):
 
-    def __init__(self, block, layers, sample_size, sample_duration, shortcut_type='B', num_classes=400):
+    def __init__(self,
+                 block,
+                 layers,
+                 sample_size,
+                 sample_duration,
+                 shortcut_type='B',
+                 num_classes=400):
         self.inplanes = 64
         super(PreActivationResNet, self).__init__()
-        self.conv1 = nn.Conv3d(3, 64, kernel_size=7, stride=(1, 2, 2),
-                               padding=(3, 3, 3), bias=False)
+        self.conv1 = nn.Conv3d(
+            3,
+            64,
+            kernel_size=7,
+            stride=(1, 2, 2),
+            padding=(3, 3, 3),
+            bias=False)
         self.bn1 = nn.BatchNorm3d(64)
         self.relu = nn.ReLU(inplace=True)
         self.maxpool = nn.MaxPool3d(kernel_size=(3, 3, 3), stride=2, padding=1)
         self.layer1 = self._make_layer(block, 64, layers[0], shortcut_type)
-        self.layer2 = self._make_layer(block, 128, layers[1], shortcut_type, stride=2)
-        self.layer3 = self._make_layer(block, 256, layers[2], shortcut_type, stride=2)
-        self.layer4 = self._make_layer(block, 512, layers[3], shortcut_type, stride=2)
-        last_duration = math.ceil(sample_duration / 16)
-        last_size = math.ceil(sample_size / 32)
-        self.avgpool = nn.AvgPool3d((last_duration, last_size, last_size), stride=1)
+        self.layer2 = self._make_layer(
+            block, 128, layers[1], shortcut_type, stride=2)
+        self.layer3 = self._make_layer(
+            block, 256, layers[2], shortcut_type, stride=2)
+        self.layer4 = self._make_layer(
+            block, 512, layers[3], shortcut_type, stride=2)
+        last_duration = int(math.ceil(sample_duration / 16))
+        last_size = int(math.ceil(sample_size / 32))
+        self.avgpool = nn.AvgPool3d(
+            (last_duration, last_size, last_size), stride=1)
         self.fc = nn.Linear(512 * block.expansion, num_classes)
 
         for m in self.modules():
@@ -129,15 +152,18 @@ class PreActivationResNet(nn.Module):
         downsample = None
         if stride != 1 or self.inplanes != planes * block.expansion:
             if shortcut_type == 'A':
-                downsample = partial(downsample_basic_block,
-                                     planes=planes * block.expansion,
-                                     stride=stride)
+                downsample = partial(
+                    downsample_basic_block,
+                    planes=planes * block.expansion,
+                    stride=stride)
             else:
                 downsample = nn.Sequential(
-                    nn.Conv3d(self.inplanes, planes * block.expansion,
-                              kernel_size=1, stride=stride, bias=False),
-                    nn.BatchNorm3d(planes * block.expansion)
-                )
+                    nn.Conv3d(
+                        self.inplanes,
+                        planes * block.expansion,
+                        kernel_size=1,
+                        stride=stride,
+                        bias=False), nn.BatchNorm3d(planes * block.expansion))
 
         layers = []
         layers.append(block(self.inplanes, planes, stride, downsample))
@@ -165,6 +191,7 @@ class PreActivationResNet(nn.Module):
 
         return x
 
+
 def get_fine_tuning_parameters(model, ft_begin_index):
     if ft_begin_index == 0:
         return model.parameters()
@@ -185,11 +212,13 @@ def get_fine_tuning_parameters(model, ft_begin_index):
 
     return parameters
 
+
 def resnet18(**kwargs):
     """Constructs a ResNet-18 model.
     """
     model = PreActivationResNet(PreActivationBasicBlock, [2, 2, 2, 2], **kwargs)
     return model
+
 
 def resnet34(**kwargs):
     """Constructs a ResNet-34 model.
@@ -204,20 +233,26 @@ def resnet50(**kwargs):
     model = PreActivationResNet(PreActivationBottleneck, [3, 4, 6, 3], **kwargs)
     return model
 
+
 def resnet101(**kwargs):
     """Constructs a ResNet-101 model.
     """
-    model = PreActivationResNet(PreActivationBottleneck, [3, 4, 23, 3], **kwargs)
+    model = PreActivationResNet(PreActivationBottleneck, [3, 4, 23, 3],
+                                **kwargs)
     return model
+
 
 def resnet152(**kwargs):
     """Constructs a ResNet-101 model.
     """
-    model = PreActivationResNet(PreActivationBottleneck, [3, 8, 36, 3], **kwargs)
+    model = PreActivationResNet(PreActivationBottleneck, [3, 8, 36, 3],
+                                **kwargs)
     return model
+
 
 def resnet200(**kwargs):
     """Constructs a ResNet-101 model.
     """
-    model = PreActivationResNet(PreActivationBottleneck, [3, 24, 36, 3], **kwargs)
+    model = PreActivationResNet(PreActivationBottleneck, [3, 24, 36, 3],
+                                **kwargs)
     return model
